Ensure that jar files works after being created
There was a problem where an error was being thrown when the jar file was being run. Mainly, file not found error.

This had to do with how the cheer resource was being retrieved. During the local implementation, cheer.txt was retrieved as a resource which was converted to a URI which was then converted to a Path object.
However, in binary files, Paths to the cheer resource do not exists. Thus, the error was thrown.

Main Implementations:

build.gradle:
1. Include the javaFxVersion in dependencies

ActionHandler.java:
1. Call the new method to that retrieves the cheer messages from resources

DataHandler.java:
1. Change the handling of the cheer resource
  a. Instead of reading saving it as a file path, implemented a method getSavedCheerMessages to read the resource as a Stream and then convert it to a List to be returned